% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{calc_gmin}
\alias{calc_gmin}
\title{Calculate Gmin}
\usage{
calc_gmin(data)
}
\arguments{
\item{\code{data}}{\code{POPBAM haplo} dataframe}
}
\description{
Gmin is the ratio of minimum between population divergence and average between population divergence. This function reads \code{POPBAM haplo} dataframe created by \code{read_popbam() and returns new dataframe with Gmin and Z-score appended to original dataframe}
}
\examples{
calc_gmin()
}
\keyword{Gmin}
\keyword{POPBAM,}

