% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{plot_popbam_tree}
\alias{plot_popbam_tree}
\title{Plot popbam trees}
\usage{
plot_popbam_tree(tree, root = FALSE)
}
\arguments{
\item{\code{tree}}{multiphylo created by \code{read_popbam_tree()}
#'}

\item{\code{root}}{name of taxon to root tree}
}
\description{
This function plots and annotates trees generated by popbam
}
\examples{
plot_popbam_trees()
}
\keyword{POPBAM,}
\keyword{plot}
\keyword{trees,}

