% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{read_popbam}
\alias{read_popbam}
\title{Read and annotate popbam output}
\usage{
read_popbam(file)
}
\arguments{
\item{\code{file}}{filename of popbam output}
}
\description{
This function reads \code{POPBAM} output (nucdiv, halpo, ld and diverge functions) and creates a appropriately labelled dataframe
}
\examples{
read_popbam()
}
\keyword{POPBAM,}
\keyword{output}
\keyword{parse}

