% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{read_popbam_trees}
\alias{read_popbam_trees}
\title{Read popbam trees}
\usage{
read_popbam_trees(x)
}
\arguments{
\item{\code{x}}{filename of popbam tree output}
}
\description{
This function reads \code{POPBAM} tree output and creates a multiphylo object with aligmnent details added to tree details
}
\examples{
read_popbam_trees()
}
\keyword{POPBAM,}
\keyword{trees}

